Move containers/discover-search to containers/layouts/discover-search
